Update sentry config in accordance to neuris recommendations
