+++ backend/src/main/resources/application-staging.yaml
@@ -15,8 +15,8 @@ sentry:
   # Add data like request headers and IP for users,
   # see https://docs.sentry.io/platforms/java/guides/spring-boot/data-management/data-collected/ for more info
   send-default-pii: true
-  # Set traces_sample_rate to 1.0 to capture 100% of transactions for tracing.
-  tracesSampleRate: 0.01
+  tracesSampleRate: 0.1 # 10% of all requests are used for samples
+  sample-rate: 1.0 # 100 % of all errors are logged
 
 frontend:
   auth:
